pyproject.toml
[project]
name = "rl-logic-seq"
version = "0.1.0"
description = "Greedy-structured offline-to-online RL for HPWL-driven logic optimization sequencing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "modal>=1.2.6",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "torch>=2.1",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
